move DrawableView and GroupPane to a new package
