--- ImageGallery/src/org/sleuthkit/autopsy/imagegallery/gui/GroupView.java
@@ -1,8 +1,0 @@
-
-package org.sleuthkit.autopsy.imagegallery.gui;
-
-
-
-public interface GroupView {
-
-}
